Clarify optional input documentation
After a bit of trial and error, I gather that the workflow API must only provide strings as input. QED, the optional inputs have to be strings. Not a big deal, except for the IDENTIFIERS value which needs to be a JSON string, which looks exactly like a YAML sequence. Updated documentation to include what I needed to see to resolve this. (Now with prettier, less-exhausting single quotes.)